Fixed a lot of shit
The code actually runs all the way to the graphics shit I don't
understand handles. I edited literally every function, and fixed so much
shit. The dynamics and the jacobians are still singular pretty much all
the time though.

THE CODE RUNS AS IS, some things may be wrong but let's keep it running
at this point!!