Fixed createDunkTrajectory and simulateBasket
I realized it doesn't matter what the value of joint five is, Nathan was
right, we just add pi/2 to whatever the value is. Other code in
simulatBasket makes it so that Ian cannot run the code, so I'm taking
that out, since once again, it doesn't matter what the value of joint
five is.